% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotslogs.R
\name{hero_builds}
\alias{hero_builds}
\title{Hero Talent Build Data}
\usage{
hero_builds(name)
}
\arguments{
\item{name}{A character vector specifying the hero name. Caution: hotslogs
will default to Abathur if a name is misspelled.}
}
\description{
This data contains information about the current top 10 most popular builds
for a particular hero.
}
\details{
The raw data is pulled from
\url{http://www.hotslogs.com/Sitewide/HeroDetails} with query string
?Hero=\code{name}.
}

