% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroesjson.R
\name{hero_data}
\alias{hero_data}
\title{Download and Parse Hero Data}
\usage{
hero_data()
}
\description{
This function downloads and parses the heroesjson data into a mangable
data.frame.
}
\details{
The raw data is pulled from \url{http://heroesjson.com/heroes.json}.
}

