% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotslogs.R
\name{hero_popularity}
\alias{hero_popularity}
\title{Hero Popularity Data}
\usage{
hero_popularity()
}
\description{
This data includes, for each hero, their name, number of games in which the
hero was played, what percentage of total games included the hero, the
percentage of the hero's games won, the change in the win rate percentage,
their role and specialty.
}
\details{
The raw data is pulled from \url{http://www.hotslogs.com/Default}.
}

