% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{heroes_data}
\alias{heroes_data}
\title{Download and Parse Heroes JSON Data}
\usage{
heroes_data()
}
\description{
This function downloads and parses the heroesjson data into a mangable
data.frame.
}
\details{
The raw data is pulled from \link{http://heroesjson.com/heroes.json}.
}

