% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotr.R
\name{hotr}
\alias{hotr}
\title{Handsontable wrapper}
\usage{
hotr(
  inputId,
  data = NULL,
  dic = NULL,
  options = NULL,
  enable_hdTypes = FALSE,
  hdTypes = NULL,
  order = NULL,
  ...
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{data}{A data frame object.}

\item{dic}{A data frame with labels of data names.}

\item{options}{A list of initialization options.}
}
\value{
Editable data
}
\description{
Handsontable wrapper for database editing,  controlling configuration and merging data
}
\examples{
\dontrun{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
   uiOutput('dstable'),
   verbatimTextOutput('test')
)
server <- function(input, output) {

output$dstable <- renderUI({
 hotr("indata1", data = mtcars, options = list(height = 300))
})

output$test <- renderPrint({
 input$indata1
})
}
shinyApp(ui, server)
}
}


}
